fix: 좋아요 등록 및 조회

+++ src/main/java/com/example/hackathon/domain/heart/dto/HeartCountResponse.java
@@ -1,4 +1,4 @@
 package com.example.hackathon.domain.heart.dto;
 
-public record HeartCountResponse (Integer heartCount) {
+public record HeartCountResponse (Long heartCount) {
 }
